Move mock contants to new file
