+++ src/constants.js
@@ -1,94 +1,8 @@
-const POSTERS = [
-  `made-for-each-other.png`,
-  `popeye-meets-sinbad.png`,
-  `sagebrush-trail.jpg`,
-  `santa-claus-conquers-the-martians.jpg`,
-  `the-dance-of-life.jpg`,
-  `the-great-flamarion.jpg`,
-  `the-man-with-the-golden-arm.jpg`,
-];
-
 const EMOJI = [
   `smile`,
   `sleeping`,
   `puke`,
   `angry`,
-];
-
-const TITLES = [
-  `The Dance of Life`,
-  `Sagebrush Trail`,
-  `The Man with the Golden Arm`,
-  `Santa Claus Conquers the Martians`,
-  `Popeye the Sailor Meets Sindbad `,
-  `The Man with the Golden Arm`,
-  `The Great Flamarion`,
-  `Santa Claus Conquers the Martians`,
-  `Made for Each Other`,
-];
-
-const SENTENCES = [
-  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-  `Cras aliquet varius magna, non porta ligula feugiat eget. `,
-  `Fusce tristique felis at fermentum pharetra.`,
-  `Aliquam id orci ut lectus varius viverra.`,
-  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-  `Aliquam erat volutpat.`,
-  `Nunc fermentum tortor ac porta dapibus.`,
-  `In rutrum ac purus sit amet tempus.`,
-];
-
-const GENRES = [
-  `Musical`,
-  `Western`,
-  `Drama`,
-  `Comedy`,
-  `Cartoon`,
-  `Mystery`,
-];
-
-const NAMES = [
-  `Luciana Khan`,
-  `Josh Cross`,
-  `Yu Fuller`,
-  `Larry Goodman`,
-  `Nida Becker`,
-  `Eshaal Carter`,
-  `Zayn Rigby`,
-  `Bill Davies`,
-  `Robson Merrill`,
-  `Dru Sharpe`,
-];
-
-const AGE_RATINGS = {
-  '0+': `General audiences`,
-  '6+': `Parental guidance`,
-  '12+': `Parents strongly cautioned`,
-  '16+': `Restricted`,
-  '18+': `No One 17 & Under Admitted`,
-};
-
-const COUNTRIES = [
-  `United States`,
-  `France`,
-  `China`,
-  `India`,
-  `United Kingdom`,
-  `Nigeria`,
-  `Egypt`,
-  `Iran`,
-  `Japan`,
-  `Korea`,
-  `Hong Kong`,
-  `Turkey`,
-  `Pakistan`,
-  `Bangladesh`,
-  `Indonesia`,
-  `Trinidad and Tobago`,
-  `Nepal`,
 ];
 
 const USER_STATUSES = [
@@ -106,22 +20,22 @@ const USER_STATUSES = [
   },
 ];
 
-const TOTAL_FILMS = 27;
+const AGE_RATINGS = {
+  '0+': `General audiences`,
+  '6+': `Parental guidance`,
+  '12+': `Parents strongly cautioned`,
+  '16+': `Restricted`,
+  '18+': `No One 17 & Under Admitted`,
+};
+
 const MAX_CARDS_TOP = 2;
 const MAX_CARDS_SHOW = 5;
 const MAX_CARDS_LOAD = 5;
 
 export {
-  POSTERS,
   EMOJI,
-  TITLES,
-  SENTENCES,
-  GENRES,
-  NAMES,
-  AGE_RATINGS,
-  COUNTRIES,
   USER_STATUSES,
-  TOTAL_FILMS,
+  AGE_RATINGS,
   MAX_CARDS_TOP,
   MAX_CARDS_SHOW,
   MAX_CARDS_LOAD,
