+++ src/main.js
@@ -1,4 +1,4 @@
-import {TOTAL_FILMS} from './constants';
+import {TOTAL_FILMS} from './mocks/constants';
 
 import Profile from './components/profile';
 import Filter from './components/filter';
